refactor: move ApiError to common-dto
