+++ common-dto/src/main/java/ru/practicum/shareit/commondto/error/ApiError.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.exception.handler;
+package ru.practicum.shareit.commondto.error;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
@@ -6,16 +6,17 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.experimental.FieldDefaults;
-import org.springframework.http.HttpStatus;
+import lombok.extern.jackson.Jacksonized;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
+@Jacksonized
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 @Getter
 public class ApiError {
-    HttpStatus status;
+    String status;
     String reason;
     String message;
